requirements.txt
Django==3.0.4
django-crispy-forms==1.9.0
djangorestframework==3.11.0